Add values for DoT and DoH transport protocols
It makes sense to be able to know whether the protocol used to
transport queries and responses is a secure one, and which one.
I kept the exact same values than the ones used by dnstap, but I
wonder if it would not make sense to support DNSCrypt UDP and TCP
as well.